requirements.txt
# Python dependencies installed after core CUDA/PyTorch stack.
# Keep torch/torchvision/torchaudio, rasterizers, nvdiffrast, and pytorch3d
# installed via the README commands for better CUDA compatibility.

einops==0.7.0
imageio[ffmpeg]==2.35.1
kornia==0.7.3
lpips==0.1.4
matplotlib==3.7.5
numpy==1.24.4
open3d==0.18.0
opencv-python==4.10.0.84
piq==0.8.0
plyfile==1.0.3
protobuf==4.25.3
pyexr==0.5.0
pytorch-msssim==1.0.0
PyYAML==6.0.2
scipy==1.10.1
tensorboard==2.14.0
tqdm==4.66.5
smplx==0.1.28
h5py==3.11.0